src/requirements.txt
Pillow==7.1.2
sanic
aiofile
